chore: move dump config function to plugin_config.go
